pyproject.toml
[project]
name = "agent-orchestrator"
version = "1.0.0"
description = "AI Agent Orchestrator with LangGraph - dynamic agent creation, multi-provider support, flexible workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.9",

    # LangGraph / LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",

    # AI Providers (direct SDKs)
    "openai>=1.50.0",
    "anthropic>=0.34.0",
    "google-generativeai>=0.8.0",

    # File Processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",

    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_orchestrator"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
